Add files via upload
Changelogs:

Added :  User prompt to enter classification type (maturity or grading) ,to be added to folder name for easy access to training log folders.
Added : test_ds.cache() , this is to save the Test Datasets generated by split_dataset function.
             test_ds.cache() will also ensure that the images on the Confusion Matrix and the actual test evaluation are the same.
Added : save_metrics_info function , this will save the metrics of the training code, including hyperparameters , test and validation results, and model summary. This is to make sure that we are able to review the train log with the modified parameters set by the user.